@@ -1,4 +1,4 @@
-﻿using OrleanSpaces.Tuples.Numerics;
+﻿using OrleanSpaces.Tuples;
 using System.Numerics;
 
 namespace OrleanSpaces.Tests;
@@ -10,8 +10,8 @@ public class Class1
     {
         if (Vector.IsHardwareAccelerated)
         {
-            IntTuple tuple = new(new[] { 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1});
-            IntTuple tuple1 = new(new[] { 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1 });
+            IntTuple tuple = new(new[] { 1, 1, 1, 1, 1, 1, 1 });
+            IntTuple tuple1 = new(new[] { 1, 1, 1, 1, 1, 1, 1 });
 
             Assert.Equal(tuple, tuple1);
         }
